# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Billionaire Analytics Terminal API — AI-powered NSE/BSE stock analytics
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StockExchange = typeof StockExchange[keyof typeof StockExchange];


export const StockExchange = {
  NSE: 'NSE',
  BSE: 'BSE',
} as const;

export type StockRecommendation = typeof StockRecommendation[keyof typeof StockRecommendation];


export const StockRecommendation = {
  BUY: 'BUY',
  HOLD: 'HOLD',
  AVOID: 'AVOID',
} as const;

export interface Stock {
  symbol: string;
  name: string;
  exchange: StockExchange;
  sector: string;
  price: number;
  previousClose: number;
  change: number;
  changePercent: number;
  dayHigh: number;
  dayLow: number;
  volume: number;
  marketCapCr: number;
  aiScore: number;
  recommendation: StockRecommendation;
  businessQualityScore: number;
  financialHealthScore: number;
  growthScore: number;
  valuationScore: number;
  riskScore: number;
  updatedAt: string;
}

export interface FinancialMetrics {
  symbol: string;
  revenueCr: number;
  revenueGrowthYoyPercent: number;
  netProfitCr: number;
  profitGrowthYoyPercent: number;
  epsInr: number;
  peRatio: number;
  pbRatio: number;
  roePercent: number;
  rocePercent: number;
  debtToEquity: number;
  dividendYieldPercent: number;
  promoterHoldingPercent: number;
}

export type NewsItemSentiment = typeof NewsItemSentiment[keyof typeof NewsItemSentiment];


export const NewsItemSentiment = {
  POSITIVE: 'POSITIVE',
  NEGATIVE: 'NEGATIVE',
  NEUTRAL: 'NEUTRAL',
} as const;

export interface NewsItem {
  id: number;
  /** @nullable */
  stockSymbol: string | null;
  headline: string;
  summary: string;
  source: string;
  sentiment: NewsItemSentiment;
  publishedAt: string;
}

export type WatchlistItemRecommendation = typeof WatchlistItemRecommendation[keyof typeof WatchlistItemRecommendation];


export const WatchlistItemRecommendation = {
  BUY: 'BUY',
  HOLD: 'HOLD',
  AVOID: 'AVOID',
} as const;

export interface WatchlistItem {
  id: number;
  symbol: string;
  name: string;
  price: number;
  changePercent: number;
  aiScore: number;
  recommendation: WatchlistItemRecommendation;
  addedAt: string;
}

export interface WatchlistInput {
  symbol: string;
}

export interface PortfolioHolding {
  id: number;
  symbol: string;
  name: string;
  quantity: number;
  avgBuyPrice: number;
  currentPrice: number;
  investedValue: number;
  currentValue: number;
  pnl: number;
  pnlPercent: number;
  buyDate: string;
}

export interface PortfolioHoldingInput {
  symbol: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  avgBuyPrice: number;
  buyDate: string;
}

export interface PortfolioHoldingUpdate {
  /** @minimum 1 */
  quantity?: number;
  /** @minimum 0 */
  avgBuyPrice?: number;
  buyDate?: string;
}

export interface PortfolioSummary {
  totalInvested: number;
  currentValue: number;
  totalPnl: number;
  totalPnlPercent: number;
  holdingsCount: number;
}

export interface DashboardSummary {
  topGainers: Stock[];
  topLosers: Stock[];
  mostActive: Stock[];
  latestNews: NewsItem[];
}

export type ListStocksParams = {
/**
 * Search term matched against symbol and name
 */
query?: string;
limit?: number;
};

export type ListNewsParams = {
limit?: number;
};