# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Billionaire Analytics Terminal API — AI-powered NSE/BSE stock analytics
 * OpenAPI spec version: 0.1.0
 */
import type { NewsItem } from './newsItem';
import type { Stock } from './stock';

export interface DashboardSummary {
  topGainers: Stock[];
  topLosers: Stock[];
  mostActive: Stock[];
  latestNews: NewsItem[];
}