# lib/api-zod/src/generated/types/financialMetrics.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Billionaire Analytics Terminal API — AI-powered NSE/BSE stock analytics
 * OpenAPI spec version: 0.1.0
 */

export interface FinancialMetrics {
  symbol: string;
  revenueCr: number;
  revenueGrowthYoyPercent: number;
  netProfitCr: number;
  profitGrowthYoyPercent: number;
  epsInr: number;
  peRatio: number;
  pbRatio: number;
  roePercent: number;
  rocePercent: number;
  debtToEquity: number;
  dividendYieldPercent: number;
  promoterHoldingPercent: number;
}