# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Billionaire Analytics Terminal API — AI-powered NSE/BSE stock analytics
 * OpenAPI spec version: 0.1.0
 */
import type { NewsItemSentiment } from './newsItemSentiment';

export interface NewsItem {
  id: number;
  /** @nullable */
  stockSymbol: string | null;
  headline: string;
  summary: string;
  source: string;
  sentiment: NewsItemSentiment;
  publishedAt: Date;
}