# lib/api-zod/src/generated/types/stock.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Billionaire Analytics Terminal API — AI-powered NSE/BSE stock analytics
 * OpenAPI spec version: 0.1.0
 */
import type { StockExchange } from './stockExchange';
import type { StockRecommendation } from './stockRecommendation';

export interface Stock {
  symbol: string;
  name: string;
  exchange: StockExchange;
  sector: string;
  price: number;
  previousClose: number;
  change: number;
  changePercent: number;
  dayHigh: number;
  dayLow: number;
  volume: number;
  marketCapCr: number;
  aiScore: number;
  recommendation: StockRecommendation;
  businessQualityScore: number;
  financialHealthScore: number;
  growthScore: number;
  valuationScore: number;
  riskScore: number;
  updatedAt: Date;
}