# lib/api-zod/src/generated/types/watchlistInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Billionaire Analytics Terminal API — AI-powered NSE/BSE stock analytics
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistInput {
  symbol: string;
}