# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Billionaire Analytics Terminal API — AI-powered NSE/BSE stock analytics
 * OpenAPI spec version: 0.1.0
 */
import type { WatchlistItemRecommendation } from './watchlistItemRecommendation';

export interface WatchlistItem {
  id: number;
  symbol: string;
  name: string;
  price: number;
  changePercent: number;
  aiScore: number;
  recommendation: WatchlistItemRecommendation;
  addedAt: Date;
}